requirements.txt
asgiref==3.8.1
Django==4.2.15
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
djangorestframework==3.15.2
pillow==10.4.0
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
